// dispenser/src/interfaces/appointment.ts
export interface IAppointment {
  CUENTA_CORRIENTE: string;
  ID_HORA_MEDICA: number;
  NUMERO_FICHA_LOCAL: number;
  PACIENTE_ID: number;
  PACIENTE_RUT: string;
  MODALIDA_ATENCION: string;
  MODALIDA_ATENCION_ID: number;
  POLICLINICO: string;
  MODULO: string;
  BOX: string;
  NOMBRE_PROFESIONAL: string;
  APELLIDO_PAT_PROFESIONAL: string;
  APELLIDO_MAT_PROFESIONAL: string;
  ID_PROFESIONAL: number;
  ATENCION_FECHA: string;
  ATENCION_HORA: string;
}